move precision checking out of normal execution path
